+++ src/components/StaticInfoTable/StaticInfoTableHeader/index.js
@@ -1,0 +1,1 @@
+export { default as StaticInfoTableHeader } from './StaticInfoTableHeader'